backend/pyproject.toml
[project]
name = "mimesis"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "websockets",
    "google-adk",
    "python-dotenv",
    "mcp",
    "httpx",
    "google-cloud-storage>=3.9.0",
    "python-multipart>=0.0.22",
    "pillow>=12.1.1",
]

[tool.ruff]
line-length = 80
